requirements.txt
Flask
flask_sqlalchemy
numpy
pandas
python_bcrypt
SQLAlchemy
gunicorn
openpyxl
fuzzywuzzy
python-Levenshtein